LayoutTests/media/video-load-readyState.html

<video controls></video>
<script src=video-test.js></script>
<script>
    function testReadyState(expected, endIt)
    {
        testExpected("video.readyState", expected);
        if (endIt)
            endTest();
    }

    testExpected("video.readyState", HTMLMediaElement.DATA_UNAVAILABLE);
    waitForEvent("loadstart", function () { testReadyState(HTMLMediaElement.DATA_UNAVAILABLE); } );
    waitForEvent("loadedmetadata", function () { testReadyState(HTMLMediaElement.DATA_UNAVAILABLE); } );
    waitForEvent("loadedfirstframe", function () { testReadyState(HTMLMediaElement.CAN_SHOW_CURRENT_FRAME); } );
    waitForEvent("canshowcurrentframe", function () { testReadyState(HTMLMediaElement.CAN_SHOW_CURRENT_FRAME); } );
    waitForEvent("canplaythrough", function () { testReadyState(HTMLMediaElement.CAN_PLAY_THROUGH, true); } );
    video.src = "content/test.mp4";
</script>
